Extract AotCompiler interface
