+++ aot/src/test/java/com/dylibso/chicory/approvals/ApprovalTest.java
@@ -1,16 +1,12 @@
 package com.dylibso.chicory.approvals;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
+import static com.dylibso.chicory.aot.AotCompiler.compileModule;
+import static com.dylibso.chicory.wasm.Parser.parse;
+import static java.lang.ClassLoader.getSystemClassLoader;
 
-import com.dylibso.chicory.aot.AotMachine;
-import com.dylibso.chicory.runtime.ExternalValues;
-import com.dylibso.chicory.runtime.HostFunction;
-import com.dylibso.chicory.runtime.Instance;
-import com.dylibso.chicory.wasm.Parser;
-import com.dylibso.chicory.wasm.types.ValueType;
-import java.io.ByteArrayOutputStream;
 import java.io.PrintWriter;
-import java.util.List;
+import java.io.StringWriter;
+import java.util.Map;
 import org.approvaltests.Approvals;
 import org.junit.jupiter.api.Test;
 import org.objectweb.asm.ClassReader;
@@ -35,14 +31,7 @@ public class ApprovalTest {
 
     @Test
     public void verifyHelloWasi() {
-        verifyGeneratedBytecode(
-                "hello-wasi.wat.wasm",
-                new HostFunction(
-                        "wasi_snapshot_preview1",
-                        "fd_write",
-                        (instance, args) -> null,
-                        List.of(ValueType.I32, ValueType.I32, ValueType.I32, ValueType.I32),
-                        List.of(ValueType.I32)));
+        verifyGeneratedBytecode("hello-wasi.wat.wasm");
     }
 
     @Test
@@ -52,17 +41,9 @@ public class ApprovalTest {
 
     @Test
     public void verifyI32Renamed() {
-        var instance =
-                Instance.builder(
-                                Parser.parse(
-                                        ClassLoader.getSystemClassLoader()
-                                                .getResourceAsStream("compiled/i32.wat.wasm")))
-                        .withMachineFactory(inst -> new AotMachine("FOO", inst))
-                        .withStart(false)
-                        .build();
-        var compiled = ((AotMachine) instance.getMachine()).compiledClass();
-
-        verifyClass(compiled);
+        var module = parse(getSystemClassLoader().getResourceAsStream("compiled/i32.wat.wasm"));
+        var result = compileModule(module, "FOO");
+        verifyClass(result.classBytes());
     }
 
     @Test
@@ -82,14 +63,7 @@ public class ApprovalTest {
 
     @Test
     public void verifyStart() {
-        verifyGeneratedBytecode(
-                "start.wat.wasm",
-                new HostFunction(
-                        "env",
-                        "gotit",
-                        (instance, args) -> null,
-                        List.of(ValueType.I32),
-                        List.of()));
+        verifyGeneratedBytecode("start.wat.wasm");
     }
 
     @Test
@@ -97,32 +71,27 @@ public class ApprovalTest {
         verifyGeneratedBytecode("trap.wat.wasm");
     }
 
-    private static void verifyGeneratedBytecode(String name, HostFunction... hostFunctions) {
-        var instance =
-                Instance.builder(
-                                Parser.parse(
-                                        ClassLoader.getSystemClassLoader()
-                                                .getResourceAsStream("compiled/" + name)))
-                        .withExternalValues(new ExternalValues(hostFunctions))
-                        .withMachineFactory(AotMachine::new)
-                        .withStart(false)
-                        .build();
-        var compiled = ((AotMachine) instance.getMachine()).compiledClass();
-
-        verifyClass(compiled);
+    private static void verifyGeneratedBytecode(String name) {
+        var module = parse(getSystemClassLoader().getResourceAsStream("compiled/" + name));
+        var result = compileModule(module);
+        verifyClass(result.classBytes());
     }
 
-    private static void verifyClass(byte[] compiled) {
-        ClassReader cr = new ClassReader(compiled);
-        var out = new ByteArrayOutputStream();
-        cr.accept(new TraceClassVisitor(new PrintWriter(out, false, UTF_8)), 0);
+    private static void verifyClass(Map<String, byte[]> classBytes) {
+        var writer = new StringWriter();
 
-        String output = out.toString(UTF_8);
+        for (byte[] bytes : classBytes.values()) {
+            ClassReader cr = new ClassReader(bytes);
+            cr.accept(new TraceClassVisitor(new PrintWriter(writer)), 0);
+            writer.append("\n");
+        }
+
+        String output = writer.toString();
         output = output.replaceAll("(?m)^ {3}FRAME.*\\n", "");
         output = output.replaceAll("(?m)^ {4}MAX(STACK|LOCALS) = \\d+\\n", "");
         output = output.replaceAll("(?m)^ *// .*\\n", "");
         output = output.replaceAll("\\n{3,}", "\n\n");
-        output = output.stripLeading();
+        output = output.strip() + "\n";
 
         Approvals.verify(output);
     }
